pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "ffmpeg-python>=0.2.0",
    "networkx>=3.5",
    "nltk>=3.9.1",
    "openai>=1.109.1",
    "plotly>=6.3.0",
    "pydub>=0.25.1",
    "python-magic>=0.4.27",
    "scikit-learn>=1.7.2",
    "streamlit>=1.50.0",
    "textblob>=0.19.0",
]
